% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{embiggen}
\alias{embiggen}
\title{Make the elements of a ggplot object larger}
\usage{
embiggen(plot = last_plot(), factor = 1.2, what = c("text", "title",
  "points", "lines"))
}
\arguments{
\item{plot}{ggplot object to embiggen}

\item{factor}{Factor by which to scale sizes}

\item{what}{What parts of the plot to embiggen?}
}
\description{
Make the elements of a ggplot object larger
}
