% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{fill_d}
\alias{fill_d}
\alias{fill_c}
\alias{color_d}
\alias{color_c}
\title{Utility wrapper for color and fill scales}
\usage{
fill_d(themename, ...)

fill_c(themename, ...)

color_d(themename, ...)

color_c(themename, ...)
}
\arguments{
\item{themename}{Name of a theme with a \code{scale_xxx_[theme]_[c/d]} function.}

\item{...}{Passed to scale function}
}
\description{
Utility wrapper for color and fill scales
}
\section{Functions}{
\itemize{
\item \code{fill_d}: discrete fill

\item \code{fill_c}: continuous fill

\item \code{color_d}: discrete color

\item \code{color_c}: continuous color
}}

