% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorclass.R
\name{hsl}
\alias{hsl}
\title{Hue Saturation Lightness color}
\usage{
hsl(h, s, l)
}
\arguments{
\item{h}{hue (in degrees), on [0, 360) (uses modulo if over)}

\item{s}{saturation, on [0, 1]}

\item{l}{lightness, on [0, 1]}
}
\description{
Hue Saturation Lightness color
}

