% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsl2rgb.R
\name{hsl2rgb}
\alias{hsl2rgb}
\title{A fixed version--colorscience does not work}
\usage{
hsl2rgb(hsl)
}
\arguments{
\item{hsl}{hue, degrees on [0, 360); saturation, on [0, 1]; lightness, on [0, 1]}
}
\value{
Matrix of RGB values on [0, 1]
}
\description{
A fixed version--colorscience does not work
}

