% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorclass.R
\name{hsv}
\alias{hsv}
\title{Hue Saturation Value color}
\usage{
hsv(h, s, v)
}
\arguments{
\item{h}{hue (in degrees), on [0, 360) (uses modulo if over)}

\item{s}{saturation, on [0, 1]}

\item{v}{value, on [0, 1]}
}
\description{
Hue Saturation Value color
}

