% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv2rgb.R
\name{hsv2rgb}
\alias{hsv2rgb}
\title{A fixed version--colorscience does not work}
\usage{
hsv2rgb(hsv)
}
\arguments{
\item{hsv}{hue, degrees on [0, 360); saturation, on [0, 1]; lightness, on [0, 1]}
}
\value{
Matrix of RGB values on [0, 1]
}
\description{
A fixed version--colorscience does not work
}

