% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorclass.R
\name{rgb}
\alias{rgb}
\title{Red Green Blue}
\usage{
rgb(r, g, b)
}
\arguments{
\item{r}{red intensity, on [0, 1]}

\item{g}{green intensity, on [0, 1]}

\item{b}{blue intensity, on [0, 1]}
}
\description{
Red Green Blue
}

