% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_emk}
\alias{theme_emk}
\alias{scale_color_emk_d}
\alias{scale_color_emk_c}
\alias{scale_fill_emk_d}
\alias{scale_fill_emk_c}
\title{ggplot theme inspired by Grandma Kangas}
\usage{
theme_emk()

scale_color_emk_d(...)

scale_color_emk_c(..., low = palette$gradient[["low"]],
  high = palette$gradient[["high"]], space = "Lab",
  guide = "colourbar")

scale_fill_emk_d(...)

scale_fill_emk_c(..., low = palette$gradient[["low"]],
  high = palette$gradient[["high"]], space = "Lab",
  guide = "colourbar")
}
\description{
ggplot theme inspired by Grandma Kangas
}
\section{Functions}{
\itemize{
\item \code{scale_color_emk_d}: discrete color scale

\item \code{scale_color_emk_c}: continuous color scale

\item \code{scale_fill_emk_d}: discrete fill scale

\item \code{scale_fill_emk_c}: continuous fill scale
}}

