% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_markdark}
\alias{theme_markdark}
\alias{theme_marklight}
\alias{scale_color_markdark_d}
\alias{scale_color_markdark_c}
\alias{scale_color_marklight_d}
\alias{scale_color_marklight_c}
\alias{scale_fill_markdark_d}
\alias{scale_fill_markdark_c}
\alias{scale_fill_marklight_d}
\alias{scale_fill_marklight_c}
\title{My own personal themes}
\usage{
theme_markdark()

theme_marklight()

scale_color_markdark_d(...)

scale_color_markdark_c(..., low = palette$gradient[["low"]],
  high = palette$gradient[["high"]], space = "Lab",
  guide = "colourbar")

scale_color_marklight_d(...)

scale_color_marklight_c(..., low = palette$gradient[["low"]],
  high = palette$gradient[["high"]], space = "Lab",
  guide = "colourbar")

scale_fill_markdark_d(...)

scale_fill_markdark_c(..., low = palette$gradient[["low"]],
  high = palette$gradient[["high"]], space = "Lab",
  guide = "colourbar")

scale_fill_marklight_d(...)

scale_fill_marklight_c(..., low = palette$gradient[["low"]],
  high = palette$gradient[["high"]], space = "Lab",
  guide = "colourbar")
}
\description{
My own personal themes
}
\section{Functions}{
\itemize{
\item \code{theme_markdark}: dark ggplot2 theme

\item \code{theme_marklight}: light ggplot2 theme

\item \code{scale_color_markdark_d}: discrete color scale

\item \code{scale_color_markdark_c}: continuous color scale

\item \code{scale_color_marklight_d}: discrete color scale

\item \code{scale_color_marklight_c}: continuous color scale

\item \code{scale_fill_markdark_d}: discrete fill scale

\item \code{scale_fill_markdark_c}: continuous fill scale

\item \code{scale_fill_marklight_d}: discrete fill scale

\item \code{scale_fill_marklight_c}: continuous fill scale
}}

