% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorclass.R
\name{yuv}
\alias{yuv}
\title{YUV color}
\usage{
yuv(y, u, v)
}
\arguments{
\item{y}{Y parameter of YUV colorspace}

\item{u}{U parameter of YUV colorspace}

\item{v}{V parameter of YUV colorspace}
}
\description{
YUV color
}

